services/data-api/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
yfinance==0.2.32
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
scipy==1.11.4
lxml==4.9.3
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
